requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
appnope
asgiref
asttokens
async-timeout
attrs
backoff
bcrypt
beautifulsoup4
bs4
build
cachetools
certifi
charset-normalizer
chroma-hnswlib
chromadb
click
coloredlogs
comm
contourpy
cycler
dataclasses-json
dateparser
debugpy
decorator
Deprecated
distro
durationpy
exceptiongroup
executing
fastapi
feedparser
fastjsonschema
filelock
firecrawl-py
flatbuffers
fonttools
frozendict
frozenlist
fsspec
google-auth
googleapis-common-protos
greenlet
groq
grpcio
h11
httpcore
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
ipykernel
ipython
jedi
Jinja2
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
kiwisolver
kubernetes
langchain
langchain-community
langchain-core
langchain-groq
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
markdown-it-py
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
mdurl
mkl-fft
mkl-random
mkl-service
mmh3
monotonic
mpmath
msgpack
multidict
multitasking
mypy-extensions
narwhals
nbformat
nest-asyncio
networkx
numpy
oauthlib
onnxruntime
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
overrides
packaging
pandas
parso
peewee
pexpect
pillow
platformdirs
plotly
posthog
prompt-toolkit
propcache
protobuf
psutil
ptyprocess
pure-eval
pyasn1
pyasn1_modules
pydantic
pydantic-settings
pydantic_core
Pygments
#pygooglenews
pyparsing
PyPika
pyproject_hooks
python-dateutil
python-dotenv
pytz
PyYAML
pyzmq
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
seaborn
sec-api
sentence-transformers
setuptools
sgmllib3k
shellingham
six
sniffio
soupsieve
SQLAlchemy
stack-data
starlette
sympy
tenacity
threadpoolctl
tokenizers
tomli
torch
tornado
tqdm
traitlets
transformers
typer
typing-inspect
typing_extensions
tzdata
tzlocal
urllib3
uvicorn
uvloop
watchfiles
wcwidth
websocket-client
websockets
wrapt
yarl
yfinance
zipp
zstandard
setuptools